Testability (#9)
* Removed some whitespace

* Renamed / moved a file.

* Starting work on the stupid AI plugin

* Wut wut

* Adds boilerplate for stupid ai systems.

* Added system that adds stupid AI to player

* Changed so that the adding component system runs in playing

* Cleanup of warnings, etc.

* Started on stupid ai movement.

* Adds new components to simplify AI decision making.

* Adds the new components to the player entity.

These will keep track of the areas a player has a presence in and
the cities the player has - to simplify a bunch of other game
functions.

This also means we could easily add tests for these new components.

* WIP WIP WIP

* When building cities a reference is added to the proper
collections in the proper component

* Yes yes

* We are getting somewhere, by adding some nice little
convenience methods on the components themselves, very nifty indeed.

* So many cool things changed.

* So many cool methods added.

* Changed name and structure of absolutely everything, bro.

Makes it easier to find stuff.

* Tests passed
everything renamed to make
it easier to find my way.

* Adds more methods for running movement tests.

* Added spec for movement that checks for all components!

* Adds spec that checks movement

* Added even more cooler code.

* So many cool and shiny things

* So many things fixed and modified making the code prettier than
ever!

* Not compiling, but we're getting there!

* OK, so removal etc is... better?

* So fancy

* Gorpadorpa

* Removed cranelift to see if it changes debugging.

Fixed conflict resolution so it works for the latest case.

* Conflicts seem to be resolved properly.

* Ooh

* Working on city support, bro

* Reworking of city support in progress.
Now the gate adds a component to every player that needs to
check city support.
The actual check can then simply check if support is needed
and then wait for player input regarding what cities to remove
etc.

* Another spec done, component added as it should.

* Fixed error in assert.

* WIP

* Woho

* Just figure some borrowing out...

* City support tests show that stuff works!

* So much good stuff done. City support checking is "done", but does it
work in the real world? Is the game playable? Nah,
it aint.

So what do we do to change that? We need a fidgin' UI, we cannot go on
without it no more.